computer_vision/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
scikit-image>=0.18.0
albumentations>=1.1.0
pycocotools>=2.0.4
tqdm>=4.62.0
tensorboard>=2.7.0
pyyaml>=5.4.1
pandas>=1.3.0
seaborn>=0.11.0
transformers>=4.20.0
accelerate>=0.20.0



